Finishing #443 (Relations) and make issues being rememberd during save, if something fails

+++ app/controllers/re_artifact_properties_controller.rb
@@ -11,6 +11,7 @@ class ReArtifactPropertiesController < RedmineReController
   end
 
   def new
+    logger.debug("NEW-Called 7777777777777777777777777777777")
     @re_artifact_properties = ReArtifactProperties.new
     @re_artifact_properties.artifact_type = params[:artifact_type].camelcase
     @artifact_type = params[:artifact_type]
@@ -50,10 +51,14 @@ class ReArtifactPropertiesController < RedmineReController
   end
 
   def create
+    logger.debug("Create-Called 7777777777777777777777777777777")
     @re_artifact_properties = ReArtifactProperties.new
     @re_artifact_properties.attributes = params[:re_artifact_properties]
     @artifact_type = @re_artifact_properties.artifact_type
 
+    @added_issue_ids = params[:issue_id]
+    @added_relations = params[:new_relation]
+    
     @bb_hash = ReBuildingBlock.find_all_bbs_and_data(@re_artifact_properties, @project.id)
     @bb_error_hash = {}
     @bb_error_hash = ReBuildingBlock.validate_building_blocks(@re_artifact_properties, @bb_error_hash, @project.id)
@@ -77,18 +82,20 @@ class ReArtifactPropertiesController < RedmineReController
 
     if @re_artifact_properties.save
       @re_artifact_properties.parent_relation.insert_at(params[:parent_relation_position])
+      handle_relations_for_new_artifact params, @re_artifact_properties.id
+      update_related_issues params
       r = :edit
     else
       logger.debug("ReArtifactProperties.create => Errors: #{@re_artifact_properties.errors.inspect}") if logger
       r = :new
     end
     initialize_tree_data
-    handle_relations params
     
     render r
   end
 
   def edit
+    logger.debug("Edit-Called 7777777777777777777777777777777")
     @re_artifact_properties = ReArtifactProperties.find(params[:id])
     @artifact_type = @re_artifact_properties.artifact_type
     @bb_hash = ReBuildingBlock.find_all_bbs_and_data(@re_artifact_properties, @project.id)
@@ -108,6 +115,7 @@ class ReArtifactPropertiesController < RedmineReController
   end
 
   def update
+    logger.debug("Update-Called 7777777777777777777777777777777")
     @re_artifact_properties = ReArtifactProperties.find(params[:id])
     @bb_hash = ReBuildingBlock.find_all_bbs_and_data(@re_artifact_properties, @project.id)
     @bb_error_hash = {}
@@ -135,12 +143,8 @@ class ReArtifactPropertiesController < RedmineReController
     end
     
     # Update related issues
-    unless params[:issue_id].blank?
-      params[:issue_id].each do |iid|
-        @re_artifact_properties.issues << Issue.find(iid)
-      end
-    end
-    
+    update_related_issues params
+        
     @artifact_type = @re_artifact_properties.artifact_type
     
     initialize_tree_data
@@ -150,6 +154,7 @@ class ReArtifactPropertiesController < RedmineReController
   end
 
   def delete
+    logger.debug("Delete-Called 7777777777777777777777777777777")
     @artifact_properties = ReArtifactProperties.find(params[:id])
 
     @relationships_incoming = @artifact_properties.relationships_as_sink
@@ -177,23 +182,53 @@ class ReArtifactPropertiesController < RedmineReController
   end
   
   def handle_relations params
-    params[:new_relation].each do |id, content|
-      if ( content['_destroy'] == "true" )
-        # id is sink id of re_artifact_properties (artifact id)
-        n = ReArtifactRelationship.find(id)
-        n.destroy
-      else
+    logger.debug("HR1-Called 7777777777777777777777777777777")
+    unless params[:new_relation].nil?
+      params[:new_relation].each do |id, content|
+        if ( content['_destroy'] == "true" )
+          # id is sink id of re_artifact_properties (artifact id)
+          n = ReArtifactRelationship.find(id)
+          n.destroy
+        else
+          # id is relation id, that should created,
+          # content contains relation_type
+          unless content['relation_type'].blank?
+            new_relation = ReArtifactRelationship.new(:source_id => params[:id], :sink_id => id, :relation_type => content['relation_type'])
+            new_relation.save
+          end 
+        end
+      end
+    end
+    
+    # If all relations are created, then this need to be cleared
+    @added_relations = nil
+  end
+  
+  def handle_relations_for_new_artifact params, new_source_artifact_id
+    logger.debug("HR2-Called 7777777777777777777777777777777")
+    unless params[:new_relation].nil?
+      params[:new_relation].each do |id, content|
         # id is relation id, that should created,
         # content contains relation_type
         unless content['relation_type'].blank?
-          new_relation = ReArtifactRelationship.new(:source_id => params[:id], :sink_id => id, :relation_type => content['relation_type'])
+          new_relation = ReArtifactRelationship.new(:source_id => new_source_artifact_id, :sink_id => id, :relation_type => content['relation_type'])
           new_relation.save
         end 
       end
     end
     
+    # If all relations are created, then this need to be cleared
+    @added_relations = nil
   end
-
+  
+  def update_related_issues params
+    unless params[:issue_id].blank?
+      params[:issue_id].each do |iid|
+        @re_artifact_properties.issues << Issue.find(iid)
+      end
+    end
+  end
+  
   def recursive_delete
     @artifact_properties = ReArtifactProperties.find(params[:id])
